pyproject.toml
[project]
name = "grpc-buf"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Soof Golan", email = "soofgolan@gmail.com" }
]
dependencies = [
    "grpcio~=1.59.3",
    "protobuf~=4.25.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
